refactor top songs results list component
