+++ src/components/Card/index.jsx
@@ -1,3 +1,5 @@
 import SectionCard from "./SectionCard";
+import TableSongCard from "./TableSongCard";
+import SongCard from "./SongCard";
 
-export { SectionCard };
+export { SectionCard, TableSongCard, SongCard };
