+++ src/pages/Home.jsx
@@ -1,4 +1,4 @@
-import { HomeSection } from "../components/Section/HomeSection";
+import { HomeSection } from "../components/Section";
 import { useRecoilValue } from "recoil";
 import { defaultAlbumsAtom, defaultPlaylistsAtom } from "../recoilState";
