+++ src/pages/Search.jsx
@@ -1,9 +1,8 @@
 import { useParams } from "react-router-dom";
 import { Col, Container, Row } from "reactstrap";
-import { SectionCard } from "../components/Card";
+import { SectionCard, SongCard } from "../components/Card";
 import { useRecoilValue } from "recoil";
 import { defaultPlaylistsAtom, searchResultsState } from "../recoilState";
-import { AiOutlineEllipsis, AiOutlineHeart } from "react-icons/ai";
 import { TopResultsSection } from "../components/Section";
 
 const Search = () => {
@@ -67,46 +66,9 @@ const Search = () => {
               <div>
                 <h2 className="mb-2 fw-bold">Songs</h2>
 
-                {[...songs]
-                  .splice(0, 4)
-                  .map(({ image_url, name, artist, duration }) => {
-                    return (
-                      <>
-                        <div className="search--songs-card d-flex align-items-center p-1 border-3">
-                          <div className="d-flex align-items-center me-auto">
-                            <img
-                              src={image_url}
-                              style={{ width: "40px", height: "40px" }}
-                            />
-                            <div className="ms-2 d-flex flex-column">
-                              <span className="p-0 m-0 fw-semibold">
-                                {name}
-                              </span>
-                              <span style={{ fontSize: "0.9rem" }}>
-                                {artist}
-                              </span>
-                            </div>
-                          </div>
-                          <div className="d-flex align-items-center me-3">
-                            <span className="me-5">
-                              <AiOutlineHeart
-                                size={20}
-                                className="search--songs-card-hidden-icon"
-                              />
-                            </span>
-                            <span className="me-2">{duration}</span>
-                            <span>
-                              <AiOutlineEllipsis
-                                size={20}
-                                fill="white"
-                                className="search--songs-card-hidden-icon"
-                              />
-                            </span>
-                          </div>
-                        </div>
-                      </>
-                    );
-                  })}
+                {[...songs].splice(0, 4).map((song) => (
+                  <SongCard key={song.id} {...song} />
+                ))}
               </div>
             </Col>
           </Row>
